Parser is now able to detect file and directory headers
There is no extracting yet as I have not figured out how I should
properly parse/store the directory structure.

There are also a few errors relating to files in my dump that seem to
have corrupted (or at least vastly different to normal) parent ID
values. I have not figured out yet if this is an issue with the firmware
dump I'm testing or its something I have to account for.